Refactor: extract platform dispatch macro and move tests to separate file
- Add define_hamming_fn! macro to generate all three platform variants
  from a single definition with shared doc comments
- Move tests to src/tests.rs
- Consolidate platform behavior docs into crate-level "Platform Behavior" section
- Simplify function doc comments to reference crate docs
- Remove timing-specific information from doc comments (kept in README)
- Fix incorrect "runtime CPU detection" claims on non-multiversion variants
